Call lua_pcall with an error handler
Currently minetest prints out the traceback after lua_pcall has returned. lua_pcall destroys the traceback before it returns. Thus all stack traces are empty. lua_pcall can take an error handler, which can itself access the traceback before lua_pcall quashes it. This error handler must be a lua object on the stack. So I took every lua_pcall without a handler (all but 1) and added a handler that just called a builtin lua function that appended the (actual) stack trace to the error.

Unfortunately the stack is also where all function arguments, lists, accumulators, return values and functions must go, so somewhere in my careful insertion of error handlers and careful cleaning up of error handlers, removing them nondestructively from the stack, the stack got trashed and the program segfaults in s_node.cpp (the one place I did NOT add error handlers to the stack).

Now, in python you would have a stack perhaps, and the error handler could be a reference completely unrelated to the stack, unable to push anything into the wrong place or accidentally delete the wrong stack entry. But we're here in lua. I spent a couple hours doing most of the groundwork here. It just needs locating where the stack gets smashed and finagling it so things work again. It's probably inserting the error handler after an item in the stack which gets used as a parameter to something, but I can't see where that might happen.

So, don't merge this but hopefully you'll find it helpful! Wish I could stop the darn thing from segfaulting.